remove debug logging; remove dead code
