+++ webapp/src/extensionManager.ts
@@ -222,7 +222,7 @@ function handleWriteCodeRequestAsync(name: string, resp: e.WriteCodeResponse, fi
             cfg.files.push(fn + ".asm");
         }
         missingDependencies?.forEach(dep => cfg.dependencies[dep] = files.dependencies[dep]);
-    }).then(() => mainPackage.saveFilesAsync(true));
+    }).then(() => mainPackage.saveFilesAsync());
 }
 
 function mkEvent(event: string): e.ExtensionEvent {
